add fg12-16ch 6.2.2

+++ Jetson_AGX_Orin_Devkits/JetPack_6.2.2_R36.5/Realsense_D4XX_JAO_FG96_8CH_Driver_JP6.2.2_R36.5/copy_d4xx_to_target.fg96.8ch.sh
@@ -85,7 +85,10 @@ if [ ! -f $max9296_dir/max9295.ko.orig ];then
 fi
 sudo cp max9295.ko $max9295_dir
 sudo cp d4xx.ko $max9295_dir
-sudo cp fzcam.ko $max9295_dir
+
+if [ -f fzcam.ko ];then
+    sudo cp fzcam.ko $max9295_dir
+fi
 
 ## copy Image file ##
 #if [ ! -f "/boot/Image.orig" ];then
